pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.11.5,<0.12.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "tzdata (>=2025.2,<2026.0)"
]

[tool.poetry]
packages = [{include = "backend_fastapi", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
run = 'fastapi dev backend_fastapi/app.py'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'


[tool.ruff]
line-length = 150


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'
